// proyecto/Entrega_2B/frontend-administracion/src/app/core/models/sample.model.ts
export interface Sample {
    id: number;
    id_function: number;
    username_user: string;
    datetime: string;
    is_augmented_data: boolean;
    data_accelerometer_x: string;
    data_accelerometer_y: string;
    data_accelerometer_z: string;
    data_gyroscope_x: string;
    data_gyroscope_y: string;
    data_gyroscope_z: string;
}